Suggestion add min css to main
when using grunt with npm or other package managers it makes it easy to pull out the appropriate distribution files. In this case the css was missing causing an auto injection to fail. Small issue not a big deal for people using CDN's or just copy and paste but for automated builds it makes tracking things down easier.